+++ InfinityHammer/InfinityHammer.cs
@@ -9,8 +9,8 @@ namespace InfinityHammer;
 [BepInPlugin(GUID, NAME, VERSION)]
 [BepInDependency("com.rolopogo.gizmo.comfy", BepInDependency.DependencyFlags.SoftDependency)]
 [BepInDependency("bruce.valheim.comfymods.gizmo", BepInDependency.DependencyFlags.SoftDependency)]
-[BepInDependency("server_devcommands", "1.83")]
-[BepInDependency("world_edit_commands", "1.63")]
+[BepInDependency("server_devcommands", "1.88")]
+[BepInDependency("world_edit_commands", "1.67")]
 public class InfinityHammer : BaseUnityPlugin
 {
   public const string GUID = "infinity_hammer";