+++ publish/CHANGELOG.md
@@ -1,4 +1,6 @@
 - v1.64
+  - Adds support for repairing MineRock objects.
+  - Improves repairing of MineRock5 objects.
   - Fixes error when removing other objects than regular build pieces.
   - Fixes remove anything not making feasts removable with the hammer.
   - Fixes `hammer_measure` giving error about birds (not measuring them anymore).